+++ crates/dialect-llvm/src/export/mod.rs
@@ -1,0 +1,93 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+//! Export LLVM dialect to textual LLVM IR.
+//!
+//! Two pieces worth knowing about live in this module: the pre-pass that
+//! assigns deterministic anonymous-value names so the textual IR is stable
+//! across runs, and the block-argument → PHI-node translation that bridges
+//! pliron's basic-block argument convention to LLVM's PHI-node convention.
+//!
+//! # Backend Configuration
+//!
+//! The export process can be customized via the [`ExportBackendConfig`] trait.
+//! Different backends (PTX, etc.) can provide their own configuration for:
+//!
+//! - Data layout string
+//! - Whether to emit `@llvm.used` for kernel preservation
+//! - Whether to emit `!nvvmir.version` metadata
+//! - Whether to emit `!nvvm.annotations` for all kernels
+//!
+//! The default [`PtxExportConfig`] uses minimal settings appropriate for standard
+//! PTX generation via llc.
+//!
+//! # Module Structure
+//!
+//! - [`config`] — backend configuration trait and built-in implementations
+//! - [`externs`] — device extern declaration types for FFI with external LTOIR
+//! - [`state`] — exporter state and kernel bookkeeping
+//! - [`names`] — value names, block labels, symbol normalization
+//! - [`module`] — module-level export flow
+//! - [`function`] — function and basic block emission
+//! - [`ops`] — operation emission
+//! - [`types`] — LLVM type printing
+//! - [`literals`] — constant/literal formatting
+//! - [`metadata`] — nvvm annotations and version, llvm.used
+
+mod config;
+mod externs;
+mod function;
+mod literals;
+mod metadata;
+mod module;
+mod names;
+mod ops;
+mod state;
+mod types;
+
+pub use config::{ExportBackendConfig, NvvmExportConfig, PtxExportConfig};
+pub use externs::{AsDeviceExtern, DeviceExternAttrs, DeviceExternDecl};
+
+use pliron::{builtin::ops::ModuleOp, context::Context};
+
+/// Export a module op to a String containing LLVM IR.
+///
+/// Uses default PTX export mode. For alternate backends, use [`export_module_to_string_with_config`].
+pub fn export_module_to_string(ctx: &Context, module: &ModuleOp) -> Result<String, String> {
+    module::export_module_to_string_with_config(ctx, module, &config::PtxExportConfig)
+}
+
+/// Export a module op with device extern declarations to a String containing LLVM IR.
+///
+/// This is the primary export function for Device FFI support. It emits:
+/// 1. Header (datalayout, target triple)
+/// 2. Device extern declarations (`declare` statements)
+/// 3. Module functions (from pliron operations)
+/// 4. Attribute groups
+/// 5. Metadata (nvvm.annotations, etc.)
+pub fn export_module_with_externs<T: AsDeviceExtern>(
+    ctx: &Context,
+    module: &ModuleOp,
+    device_externs: &[T],
+    config: &dyn ExportBackendConfig,
+) -> Result<String, String> {
+    let externs: Vec<DeviceExternDecl> = device_externs
+        .iter()
+        .map(|e| e.as_device_extern())
+        .collect();
+    module::export_module_with_externs_impl(ctx, module, &externs, config)
+}
+
+/// Export a module op to a String containing LLVM IR with custom backend configuration.
+///
+/// The `config` parameter controls backend-specific IR generation options like
+/// data layout, metadata emission, and symbol preservation.
+pub fn export_module_to_string_with_config(
+    ctx: &Context,
+    module: &ModuleOp,
+    config: &dyn ExportBackendConfig,
+) -> Result<String, String> {
+    module::export_module_to_string_with_config(ctx, module, config)
+}
